requirements.txt
alembic==1.7.6
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-pytest==0.0.5
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
numpy==1.22.1
packaging==21.3
pandas==1.4.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-asyncio==0.17.2
python-dateutil==2.8.2
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2021.3
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.2
